+++ littlefs-rust-core/src/fs/traverse.rs
@@ -205,12 +205,10 @@ pub fn lfs_fs_traverse_(
                 lfs_pair_fromle32(&mut raw);
 
                 let tag = tag.unwrap();
-                if u32::from(lfs_tag_type3(tag as u32)) == LFS_TYPE_CTZSTRUCT {
+                if u32::from(lfs_tag_type3(tag)) == LFS_TYPE_CTZSTRUCT {
                     let lfs_rcache = borrow_unchecked(&mut lfs.rcache);
                     lfs_ctz_traverse(lfs, None, lfs_rcache, raw[0], raw[1], Some(cb), data)?;
-                } else if includeorphans
-                    && u32::from(lfs_tag_type3(tag as u32)) == LFS_TYPE_DIRSTRUCT
-                {
+                } else if includeorphans && u32::from(lfs_tag_type3(tag)) == LFS_TYPE_DIRSTRUCT {
                     #[allow(clippy::needless_range_loop)] // Rule 2: preserve C loop structure
                     for i in 0..2 {
                         cb(data, raw[i])?;
